Improving speed and output info for build procedure.

+++ build.sh
@@ -2,27 +2,35 @@
 
 # Assumes PHP CLI, curl, and zip are already installed.
 
-# install NVM, Node, and NPM
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/master/install.sh | bash
-export NVM_DIR="$HOME/.nvm"
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
-nvm install node
+# check if node is installed and install it if not
+if ! command -v node &> /dev/null; then
+  echo "Node.js is not installed. Installing Node.js and NPM using NVM..."
 
-# update version in various json files
+  # install NVM, Node, and NPM
+  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/master/install.sh | bash
+  export NVM_DIR="$HOME/.nvm"
+  [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
+#  [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
+  nvm install node
+fi
+
+echo "Update version in various json files..."
 node ./buildPipeline/versionUpdate.js
 
-# update NPM packages
-npm update
+echo "Update NPM packages..."
+npm update --no-audit --prefer-offline
 
+echo "Cleaning up old build directory..."
 rm -r build
-rm touchpoint-wp.zip
+if [ -f touchpoint-wp.zip ]; then
+  rm touchpoint-wp.zip
+fi
 mkdir build
 mkdir build/assets
 mkdir build/assets/js
 
-# install uglify and uglify the JS files.
-echo $(npm install -g uglify-js)
+echo "Install uglify and uglify the JS files..."
+#echo $(npm install -g uglify-js)
 uglifyjs assets/js/base-defer.js -o build/assets/js/base-defer.min.js --source-map
 uglifyjs assets/js/meeting-defer.js -o build/assets/js/meeting-defer.min.js --source-map
 uglifyjs assets/js/partner-defer.js -o build/assets/js/partner-defer.min.js --source-map
@@ -30,10 +38,12 @@ cp -r assets build
 cd ./build || exit
 cd ..
 
-# build blocks
-npm install -g @wordpress/scripts
-wp-scripts build --webpack-src-dir=blocks --output-path=build/blocks
-wp-scripts build-blocks-manifest --input=blocks --output=build/blocks/blocks-manifest.php
+echo "Build blocks..."
+#npm install -g @wordpress/scripts
+npx wp-scripts build --webpack-src-dir=blocks --output-path=build/blocks
+npx wp-scripts build-blocks-manifest --input=blocks --output=build/blocks/blocks-manifest.php
+
+echo "Internationalization..."
 
 cp -r ./i18n ./build/i18n
 
@@ -41,6 +51,8 @@ php ./wp-cli.phar i18n make-json ./build/i18n
 php ./wp-cli.phar i18n make-mo ./build/i18n
 cp ./wpml-config.xml ./build/wpml-config.xml
 cp ./composer.json ./build/composer.json
+
+echo "Copy stuff..."
 
 cp -r ./ext ./build/ext
 cp -r ./src ./build/src
